commit comments feature, Chapter 4.10

+++ models/posts.js
@@ -1,6 +1,6 @@
 const marked = require('marked')
 const Post = require('../lib/mongo').Post
-
+const CommentModel = require('./comments')
 // 将 post 的 content 从 markdown 转换成 html
 Post.plugin('contentToHtml',{
   afterFind:function (posts){
@@ -16,6 +16,26 @@ Post.plugin('contentToHtml',{
     return post
   }
 })
+// 给 post 添加留言数 commentsCount
+Post.plugin('addCommentsCount',{
+  afterFind:function(posts){
+    return Promise.all(posts.map(function (post){
+      return CommentModel.getCommentCount(post._id).then(function (commentsCount){
+        post.commentsCount = commentsCount
+        return post
+      })
+    }))
+  },
+  afterFindOne:function (post){
+    if(post){
+      return CommentModel.getCommentCount(post._Id).then(function (count){
+        post.commentsCount = count
+        return post
+      })
+    }
+    return post
+  }
+})
 module.exports = {
   //创建一篇文章
   create:function create(post){
@@ -27,6 +47,7 @@ module.exports = {
       .findOne({ _id: postId })
       .populate({ path: 'author', model: 'User' })
       .addCreatedAt()
+      .addCommentsCount()
       .contentToHtml()
       .exec()
   },
@@ -42,6 +63,7 @@ module.exports = {
         .populate({ path: 'author', model: 'User' })
         .sort({ _id: -1 })
         .addCreatedAt()
+        .addCommentsCount()
         .contentToHtml()
         .exec()
     },
@@ -64,7 +86,13 @@ module.exports = {
     return Post.update({_id:postId},{$set:data}).exec()
   },
   // 通过文章 id 删除一篇文章
-  delPostById:function delPostById(postId){
-    return Post.deleteOne({_id:postId}).exec()
+  delPostById:function delPostById(postId,author){
+    return Post.deleteOne({author:author,_id:postId}).exec()
+      .then(function (res){
+        // 文章删除后，再删除该文章下的所有留言
+        if(res.result.ok && res.result.n>0){
+          return CommentModel.delCommentByPostId(postId)
+        }
+      })
   }
 }
